requirements.txt
pandas==1.4.2
pymongo==4.1.1
pandas_datareader==0.10.0
yfinance==0.1.70
statsmodels==0.13.2
numpy==1.22.3
openpyxl==3.0.9
pip==23.3.2
plotly-express
kaleido
pillow
matplotlib
xlsxwriter